+++ R/simmer-class.R
@@ -212,9 +212,13 @@ Simmer <- R6Class("simmer",
 
     get_n_generated = function(source) get_n_generated_(private$sim_obj, source),
 
+    get_trajectory = function(source) get_trajectory_(private$sim_obj, source)[],
+
     get_name = function() get_name_(private$sim_obj),
 
-    get_attribute = function(keys, global=FALSE) get_attribute_(private$sim_obj, keys, global),
+    get_attribute = function(keys) get_attribute_(private$sim_obj, keys, FALSE),
+
+    get_global = function(keys) get_attribute_(private$sim_obj, keys, TRUE),
 
     get_prioritization = function() get_prioritization_(private$sim_obj),
 
@@ -258,7 +262,6 @@ Simmer <- R6Class("simmer",
       )
     },
 
-    # not exposed, internal use
     get_sources = function() { private$src },
     get_resources = function() { private$res }
   ),