+++ tests/testthat/test-simmer-dataframe.R
@@ -78,13 +78,17 @@ test_that("absolute time works as expected", {
 })
 
 test_that("generates the expected amount", {
+  t <- trajectory() %>% timeout(0)
   env <- simmer(verbose = TRUE) %>%
-    add_dataframe("dummy", trajectory(), data.frame(time=rep(1, 3))) %>%
+    add_dataframe("dummy", t, data.frame(time=rep(1, 3))) %>%
     run()
   arr <- get_mon_arrivals(env)
 
+  expect_equal(env %>% get_sources(), "dummy")
   expect_error(env %>% get_n_generated("asdf"))
   expect_equal(env %>% get_n_generated("dummy"), 3)
+  expect_error(env %>% get_trajectory("asdf"))
+  expect_equal(env %>% get_trajectory("dummy"), t)
   expect_equal(arr$start_time, 1:3)
   expect_equal(arr$end_time, 1:3)
   expect_equal(arr$activity_time, rep(0, 3))